# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: EarnQA API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /questions/{id}:
    get:
      operationId: getQuestion
      tags: [questions]
      summary: Get a question by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Question details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuestionWithStats"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /wallet/withdraw:
    post:
      operationId: requestWithdrawal
      tags: [wallet]
      summary: Request a withdrawal
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WithdrawalRequestBody"
      responses:
        "201":
          description: Withdrawal requested
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
        "400":
          description: Bad request or insufficient funds
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /users/me:
    get:
      operationId: getMe
      tags: [users]
      summary: Get current user profile
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
    patch:
      operationId: updateMe
      tags: [users]
      summary: Update current user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProfileBody"
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    Question:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        type:
          type: string
          enum: [short_answer, poll, rating]
        category:
          type: string
        status:
          type: string
          enum: [active, pending, rejected]
        pollOptions:
          type: ["array", "null"]
          items:
            type: string
        creatorId:
          type: ["string", "null"]
        creatorName:
          type: ["string", "null"]
        isCustom:
          type: boolean
        totalAnswers:
          type: integer
        rejectionReason:
          type: ["string", "null"]
        rejectedAt:
          type: ["string", "null"]
        rejectedBy:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - title
        - type
        - category
        - status
        - isCustom
        - totalAnswers
        - createdAt
    QuestionWithStats:
      allOf:
        - $ref: "#/components/schemas/Question"
        - type: object
          properties:
            ratingAverage:
              type: ["number", "null"]
            pollResults:
              type: ["array", "null"]
              items:
                type: object
                properties:
                  option:
                    type: string
                  count:
                    type: integer
                  percentage:
                    type: number
                required:
                  - option
                  - count
                  - percentage
            userHasAnswered:
              type: boolean
          required:
            - userHasAnswered
...
    Transaction:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: string
        type:
          type: string
          enum: [earning, withdrawal, question_creation, creator_bonus, creator_reward, referral_signup_bonus, referral_answer_bonus, referral_reversal, profile_reward]
        amountCents:
          type: number
        description:
          type: string
        status:
          type: string
          enum: [completed, pending, rejected, approved, transferred]
        accountTitle:
          type: string
          nullable: true
        bankName:
          type: string
          nullable: true
        approvedAt:
          type: string
          nullable: true
        transferredAt:
          type: string
          nullable: true
        createdAt:
          type: string
      required:
        - id
        - userId
        - type
        - amountCents
        - description
        - status
        - createdAt
...
    WithdrawalRequestBody:
      type: object
      properties:
        amountCents:
          type: integer
        paymentMethod:
          type: string
        paymentDetails:
          type: string
        accountTitle:
          type: string
        bankName:
          type: string
      required:
        - amountCents
        - paymentMethod
        - paymentDetails
    UserProfile:
      type: object
      properties:
        id:
          type: integer
        clerkId:
          type: string
        email:
          type: string
        name:
          type: ["string", "null"]
        city:
          type: ["string", "null"]
        ageGroup:
          type: ["string", "null"]
        gender:
          type: ["string", "null"]
        isAdmin:
          type: boolean
        balanceCents:
          type: number
        questionCount:
          type: number
        answerCount:
          type: number
        earningsCents:
          type: number
        referralCode:
          type: ["string", "null"]
        referredByUserId:
          type: ["string", "null"]
        isNew:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - clerkId
        - email
        - isAdmin
        - isNew
        - createdAt
    UpdateProfileBody:
      type: object
      properties:
        name:
          type: ["string", "null"]
        city:
          type: ["string", "null"]
        ageGroup:
          type: ["string", "null"]
        gender:
          type: ["string", "null"]
...